+++ src/ska_mid_cbf_fhs_vcc/vcc_all_bands/vcc_all_bands_device.py
@@ -157,7 +157,7 @@ class VCCAllBandsController(FhsControllerBaseDevice[VCCAllBandsComponentManager]
             "or a value per frequency slice to be applied separately."
         ),
     )
-    def AutoSetFilterGains(self, headroom: list[float] = [3.0]) -> DevVarLongStringArrayType:
+    def AutoSetFilterGains(self: VCCAllBandsController, headroom: list[float] | None = None) -> DevVarLongStringArrayType:
         """Tango command to start a scan operation.
 
         Args:
@@ -169,6 +169,8 @@ class VCCAllBandsController(FhsControllerBaseDevice[VCCAllBandsComponentManager]
             :obj:`tuple[list[ResultCode], list[str]]`: The Tango result code and a string
             message indicating status. The message is for information purpose only.
         """
+        if headroom is None:
+            headroom = [3.0]
         command_handler = self.get_command_object(command_name="AutoSetFilterGains")
         # It is important that the argin keyword be provided, as the
         # component manager method will be overriden in simulation mode